Updated build process to use MSVC

+++ include/calc.h
@@ -8,6 +8,6 @@
 LRESULT CALLBACK WndProc(HWND hwnd, UINT msg, WPARAM wParam, LPARAM lParam);
 void calc_render_gui(HWND hwnd, HINSTANCE hInstance);
 void calc_process_commands(HWND hwnd, WORD wParam);
-int calc_lookup_button_id_from_key_id(int search_key_id);
+WORD calc_lookup_button_id_from_key_id(int search_key_id);
 
 #endif /* CALC_H */
